Moved OrganisationPopulator into Goodcity namespace.
