+++ lib/tasks/populate_organisation.rake
@@ -1,7 +1,9 @@
 #use 'rake populate_organisation:organisation' to create or update organisation details
 
+require "goodcity/organisation_populator"
+
 namespace :populate_organisation do
   task organisation:  :environment do
-    OrganisationPopulator.new().populate_organisation_db
+    Goodcity::OrganisationPopulator.new.populate_organisation_db
   end
 end
